Moved parser files in grammar_compiler/parser
