+++ src/grammar_compiler/__init__.py
@@ -1,15 +1,6 @@
-from .binaryfile import BinaryFile
-from .parser import Parser
-from .any_reader import AnyReader
-from .all_reader import AllReader
-from .character_reader import CharacterReader
+from . import parser
 
 
 __all__ = [
-    'BinaryFile',
-    'Parser',
-    'CharacterReader',
-    'AnyReader',
-    'AllReader',
-    'CharacterReader'
+    'parser'
 ]
