+++ src/grammar_compiler/parser/all_reader.py
