+++ src/grammar_compiler/parser/binaryfile.py
