+++ src/test_file.py
@@ -1,5 +1,5 @@
 import os
-from .grammar_compiler import BinaryFile
+from .grammar_compiler.parser import BinaryFile
 
 
 test_file_name = "test_file.txt"
